move reader -> old_reader
